# UserAccountService/src/main/resources/swagger.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:7676'
basePath: /
tags:
  - name: user-account-controller
    description: User Account Controller
  - name: User Account Service
    description: User accounts
paths:
  /useraccount/authenticate:
    post:
      tags:
        - user-account-controller
      summary: authenticate User Account
      operationId: authenticate
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: userDetails
          description: userDetails
          required: true
          schema:
            $ref: '#/definitions/UserDetails'
      responses:
        '200':
          description: OK
          schema:
            type: string
  /useraccount/create:
    post:
      tags:
        - user-account-controller
      summary: Create a User Account
      operationId: create
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: userDetails
          description: userDetails
          required: true
          schema:
            $ref: '#/definitions/UserDetails'
      responses:
        '200':
          description: OK
          schema:
            type: string
        '226':
          description: The username already exists
          schema:
            $ref: '#/definitions/UserAccountError'
  '/useraccount/delete/{userID}':
    delete:
      tags:
        - user-account-controller
      summary: Delete a User Account
      operationId: delete
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: userID
          in: path
          description: userID
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: string
  '/useraccount/enable/{userID}':
    patch:
      tags:
        - user-account-controller
      summary: Enable a User Account
      operationId: enable
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: userID
          in: path
          description: userID
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: string
  /useraccount/users:
    get:
      tags:
        - user-account-controller
      summary: Fetch User Accounts
      operationId: users
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/AzureUser'
definitions:
  AzureUser:
    type: object
    properties:
      accountEnabled:
        type: boolean
      displayName:
        type: string
      id:
        type: string
      mailNickname:
        type: string
      passwordProfile:
        $ref: '#/definitions/PasswordProfile'
      userPrincipalName:
        type: string
  PasswordProfile:
    type: object
    properties:
      forceChangePasswordNextSignIn:
        type: boolean
      password:
        type: string
  UserAccountError:
    type: object
    properties:
      errors:
        type: array
        items:
          type: string
      message:
        type: string
      status:
        type: string
        enum:
          - '100'
          - '101'
          - '102'
          - '103'
          - '200'
          - '201'
          - '202'
          - '203'
          - '204'
          - '205'
          - '206'
          - '207'
          - '208'
          - '226'
          - '300'
          - '301'
          - '302'
          - '303'
          - '304'
          - '305'
          - '307'
          - '308'
          - '400'
          - '401'
          - '402'
          - '403'
          - '404'
          - '405'
          - '406'
          - '407'
          - '408'
          - '409'
          - '410'
          - '411'
          - '412'
          - '413'
          - '414'
          - '415'
          - '416'
          - '417'
          - '418'
          - '419'
          - '420'
          - '421'
          - '422'
          - '423'
          - '424'
          - '426'
          - '428'
          - '429'
          - '431'
          - '451'
          - '500'
          - '501'
          - '502'
          - '503'
          - '504'
          - '505'
          - '506'
          - '507'
          - '508'
          - '509'
          - '510'
          - '511'
  UserDetails:
    type: object
    properties:
      emailAddress:
        type: string
      password:
        type: string